requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Dependencias para módulo Socios de Negocio
beautifulsoup4==4.12.2
requests==2.31.0
pillow==10.1.0
pytesseract==0.3.10

# Dependencias para módulo System Config
pytz==2023.3
python-dateutil==2.8.2
